delete category fix,  bcrypt update

+++ src/ExpenseTrackerApi.Application/Categories/Commands/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -25,20 +25,13 @@ public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryComman
             .AsNoTrackingWithIdentityResolution()
             .Where(x => x.CategoryId == category.Id)
             .ToListAsync(cancellationToken);
+        if (budgets.Count > 0) throw new Exception("Category has budgets!");
+
         var expenses = await _dbContext.Expenses
             .AsNoTrackingWithIdentityResolution()
             .Where(x => x.CategoryId == category.Id)
             .ToListAsync(cancellationToken);
-        foreach (var budget in budgets)
-        {
-            _dbContext.Budgets.Attach(budget);
-            _dbContext.Budgets.Remove(budget);
-        }
-        foreach (var expense in expenses)
-        {
-            _dbContext.Expenses.Attach(expense);
-            _dbContext.Expenses.Remove(expense);
-        }
+        if (expenses.Count > 0) throw new Exception("Category has expenses!");
 
         _dbContext.Categories.Attach(category);
         _dbContext.Categories.Remove(category);
